refactor: stop using the legacy standalone server
