+++ .changeset/many-moles-sing.md
@@ -1,0 +1,5 @@
+---
+'@backstage/backend-common': patch
+---
+
+Deprecate the legacy `createServiceBuilder`.
