+++ plugins/app-backend/package.json
@@ -66,6 +66,7 @@
   },
   "devDependencies": {
     "@backstage/backend-app-api": "workspace:^",
+    "@backstage/backend-defaults": "workspace:^",
     "@backstage/backend-test-utils": "workspace:^",
     "@backstage/cli": "workspace:^",
     "@backstage/types": "workspace:^",
