--- plugins/app-backend/src/service/standaloneServer.ts
@@ -1,51 +1,0 @@
-/*
- * Copyright 2020 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { Server } from 'http';
-import { createServiceBuilder } from '@backstage/backend-common';
-import { Config } from '@backstage/config';
-import { createRouter } from './router';
-import { LoggerService } from '@backstage/backend-plugin-api';
-
-export interface ServerOptions {
-  port: number;
-  enableCors: boolean;
-  config: Config;
-  logger: LoggerService;
-}
-
-export async function startStandaloneServer(
-  options: ServerOptions,
-): Promise<Server> {
-  const logger = options.logger.child({ service: 'app-backend' });
-  logger.debug('Starting application server...');
-  const router = await createRouter({
-    logger,
-    config: options.config,
-    appPackageName: 'example-app',
-  });
-
-  const service = createServiceBuilder(module)
-    .setPort(options.port)
-    .addRouter('', router);
-
-  return await service.start().catch(err => {
-    logger.error(err);
-    process.exit(1);
-  });
-}
-
-module.hot?.accept();
