+++ plugins/catalog-backend/package.json
@@ -91,6 +91,7 @@
     "zod": "^3.22.4"
   },
   "devDependencies": {
+    "@backstage/backend-defaults": "workspace:^",
     "@backstage/backend-test-utils": "workspace:^",
     "@backstage/cli": "workspace:^",
     "@backstage/plugin-permission-common": "workspace:^",
